hello world
First bit of c++ code. A digital watchface that shows the time in either 24 or 12 hour formats (config variable added, set to 12 or 24) with the date and step count. step count doesn't reset daily, so that's a thing to work on at some point. also started working on an idea for a simple calculator.