Added OswAppSwitcher
The AppSwitcher is working fine.
I added the Option for AutoSleep and the DeepSleep Method from overlays to the App itself.
I setup the main.cpp to work with AppSwitcher, but did not clean it up because i am not sure if that is the ways how it will be done :)