config value seems to have been removed - replacing
Until we can figure out a way to get the time format to update on-device, we probably need to keep this in config. I experimented with adding it to the manual time config screen, but when changing the time format, it causes the screen to go blank and if you press the home button, it doesn't seem to reflect the updated time format. unsure what needs to be done to get the os to reflect the updated config value