[untested] sensitivity for raise to wake
I think this might be what we need to make the sensitivity adjustable, but I haven't been able to experiment/test with my watch being out of commission for a bit. In the mean time, maybe someone who knows more about the hardware etc would be able to identify if this is on the right track or completely wrong. this is related to https://github.com/Open-Smartwatch/open-smartwatch-os/issues/99