Allowing on-device brightness updates
Adding new config value - minimum brightness. This is to prevent the user from setting the brightness to 0 on accident and forgetting about it. The minimum can be changed to 0 if the user would like to be able to set the brightness to 0. This change persists the brightness the user selects on-device. Might be nice to introduce some kind of helper method or something in OswConfig that handles enableWrite and disableWrite for reducing repetitive code.