Revert "workflows: fix code style workflows user name (#410)"
Because this will assign potentially a lot of additional code changes to only the pipeline triggerer.
Example: Multiple deves are working on the same PR. One of them merges it in the end (or better
a maintainer of us) and "commits" all the changes. This will cause in Git blame to only show him up.
Also, filtering based on the actions email will not work anymore.
Furthermore, people signing commits (like me) will cause a lot of unverified commits to pop up.

This reverts commit 22e3bfa614fa50c327e98a878513b54282683ec5.