emulator: build fixes for clang
With Xcode 26.2 17C52 and Apple clang version 17.0.0 (clang-1700.6.3.2),
these changes were necessary to overcome build errors.

…/emulator/src/tests/unitTests/logging.cpp:35:73: error: no matching literal operator for call to 'operator""d' with argument of type 'long double' or 'const char *', and no matching literal operator template
   35 |     OSW_LOG_D(String("String"), "c_str", 42, 'a', false, 42.042f, 42.042d);
      |

42.042d is an invalid literal. The intent is to interpret the literal as
a double, and for this no suffix is required in standard C++. The “d”
literal suffix is a GCC extension. The code is equally correct without
the “d” suffix.

…/emulator/src/tests/unitTests/string.cpp:190:47: error: use of overloaded operator '+' is ambiguous (with operand types 'char' and 'String')
  190 |     EXPECT_STREQ("ac420c42.424200", (String(a + c) + b + c + d + e).c_str()); // a + c will be processed by a std::string instead of String!
      |                                             ~ ^ ~

The comment clarifies that it is intended for this to be handled by
std::string, so construct a std::string manually to resolve the
ambiguity.